Lots of changes!
Renamed the header file for the Parser_Context class.
Added a file for the implementation of the Parser_Context class.
Added a strings pool to this class.
Added the lexer rules to read numeric inputs.
Added the lexer rules to read string inputs.
There are some testing rules in the parser.